Qsine/web-scraper/requirements.txt
requests
bs4
openai
Pillow
flask
selenium
webdriver-manager
pymongo
aiohttp